Catch location not found error location search
It is possible that the `GoogleV3.geocode(...)` returns None. If this happens it means no location was found and we should display a proper error message to the User. For this we check if it is None and raise the `GeocoderQueryError` error, which is already handled to give the user the correct error. That `None` is returned may be due to a change in the Google Maps API.